Extract advertised endpoint helpers into shared package
- Move advertisedEndpoint utilities from client-runtime to shared
- Re-export shared helpers for existing consumers
- Add tests for URL normalization and endpoint construction
