+++ packages/shared/package.json
@@ -7,6 +7,10 @@
       "types": "./src/model.ts",
       "import": "./src/model.ts"
     },
+    "./advertisedEndpoint": {
+      "types": "./src/advertisedEndpoint.ts",
+      "import": "./src/advertisedEndpoint.ts"
+    },
     "./git": {
       "types": "./src/git.ts",
       "import": "./src/git.ts"
